+++ src/base/b-image/b-image.ts
@@ -14,7 +14,7 @@ import { getSrcSet } from 'core/html';
 import iProgress from 'traits/i-progress/i-progress';
 import iVisible from 'traits/i-visible/i-visible';
 
-import iBlock, { component, prop, wait, hook, ModsDecl } from 'super/i-block/i-block';
+import iBlock, { component, prop, hook, wait, ModsDecl } from 'super/i-block/i-block';
 export * from 'super/i-block/i-block';
 
 export type SizeType =
@@ -27,41 +27,41 @@ export const
 @component({functional: {}})
 export default class bImage extends iBlock implements iProgress, iVisible {
 	/**
-	 * Target image src (fallback if srcset provided)
+	 * Image src (fallback if srcset provided)
 	 */
 	@prop({
 		type: String,
-		watch: {fn: 'initOverlay', immediate: true}
+		watch: {fn: 'init', immediate: true}
 	})
 
 	readonly src: string = '';
 
 	/**
-	 * Target images srcset
+	 * Image "srcset" attribute
 	 */
 	@prop({type: Object, required: false})
 	readonly srcset?: Dictionary<string>;
 
 	/**
-	 * Target sizes HTML attribute
+	 * Image "sizes" attribute
 	 */
 	@prop({type: String, required: false})
 	readonly sizes?: string;
 
 	/**
-	 * Alternate text
+	 * Alternate text for the image
 	 */
 	@prop({type: String, required: false})
 	readonly alt?: string;
 
 	/**
-	 * Background size type
+	 * Image background size type
 	 */
 	@prop(String)
 	readonly sizeType: SizeType = 'contain';
 
 	/**
-	 * Image position
+	 * Image background position
 	 */
 	@prop(String)
 	readonly position: string = '50% 50%';
@@ -85,13 +85,15 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 	readonly afterImg?: CanArray<string>;
 
 	/**
-	 * Parameters for a overlay image
+	 * Parameters for an overlay image
+	 * (when the image is loading)
 	 */
 	@prop({type: [String, Object], required: false})
 	readonly overlayImg?: string | Dictionary;
 
 	/**
 	 * Parameters for a broken image
+	 * (when the image loading was failed)
 	 */
 	@prop({type: [String, Object], required: false})
 	readonly brokenImg?: string | Dictionary;
@@ -115,14 +117,14 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 	/**
 	 * Initializes an image loading process
 	 */
-	@wait('ready', {label: $$.initOverlay})
-	protected initOverlay(): CanPromise<void> {
+	@wait('ready', {label: $$.init})
+	protected init(): CanPromise<void> {
 		const
 			tmpSrc = <CanUndef<string>>this.tmp[this.src];
 
 		if (tmpSrc) {
 			this.updateHeight(tmpSrc);
-			this.onImageLoaded(tmpSrc);
+			this.onImageLoadSuccess(tmpSrc);
 			return;
 		}
 
@@ -147,11 +149,11 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 
 		this.async
 			.promise(img.init, {label: $$.loadImage})
-			.then(() => this.onImageLoaded(img), this.onError);
+			.then(() => this.onImageLoadSuccess(img), this.onImageLoadFailed);
 	}
 
 	/**
-	 * Sets an image's height according to its ratio
+	 * Updates an image height according to its ratio
 	 * @param img
 	 */
 	protected updateHeight(img: HTMLImageElement | string): void {
@@ -166,7 +168,14 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 				tmpPadding = `${(1 / this.ratio) * 100}%`;
 
 			} else if (!Object.isString(img) && this.ratio !== 0) {
-				tmpPadding = `${(1 / this.computeRatio(img)) * 100}%`;
+				let
+					ratio = 1;
+
+				if (img && (img.naturalHeight || img.naturalWidth)) {
+					ratio = img.naturalHeight === 0 ? 1 : img.naturalWidth / img.naturalHeight;
+				}
+
+				tmpPadding = `${(1 / ratio) * 100}%`;
 
 			} else {
 				tmpPadding = '';
@@ -180,26 +189,11 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 	}
 
 	/**
-	 * Calculates the specified image aspect ratio
-	 * @param img
-	 */
-	protected computeRatio(img: CanUndef<HTMLImageElement>): number {
-		const
-			{naturalHeight = 0, naturalWidth = 0} = img || {};
-
-		if (naturalHeight || naturalWidth) {
-			return naturalHeight === 0 ? 1 : naturalWidth / naturalHeight;
-		}
-
-		return 1;
-	}
-
-	/**
-	 * Saves image content within a cache and destroys it
+	 * Saves an image style to the cache
 	 */
 	@hook('beforeDestroy')
 	@wait('loading')
-	protected destroyImage(): CanPromise<void> {
+	protected memoizeImage(): CanPromise<void> {
 		const
 			{img} = this.$refs;
 
@@ -207,8 +201,6 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 			this.tmp[this.src] = img[$$.img];
 			this.tmp[`${this.src}-padding`] = img.style.paddingBottom;
 		}
-
-		img.style.backgroundImage = '';
 	}
 
 	/** @override */
@@ -218,12 +210,12 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 	}
 
 	/**
-	 * Handler: image loaded
+	 * Handler: image loading successfully completed
 	 *
 	 * @param img
 	 * @emits loadSuccess()
 	 */
-	protected onImageLoaded(img: HTMLImageElement | string): void {
+	protected onImageLoadSuccess(img: HTMLImageElement | string): void {
 		const
 			{img: imgRef} = this.$refs,
 			cssImg = Object.isString(img) ? img : `url("${img.currentSrc}")`;
@@ -242,12 +234,12 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 	}
 
 	/**
-	 * Handler: image load error
+	 * Handler: image loading failed
 	 *
 	 * @param err
 	 * @emits loadFail()
 	 */
-	protected onError(err: CanUndef<Error | AsyncCtx>): void {
+	protected onImageLoadFailed(err: CanUndef<Error | AsyncCtx>): void {
 		this.setMod('progress', false);
 
 		if (err && 'type' in err && err.type === 'clearAsync') {
@@ -256,5 +248,12 @@ export default class bImage extends iBlock implements iProgress, iVisible {
 
 		this.setMod('showError', true);
 		this.emitError('loadFail', err);
+	}
+
+	/** @override */
+	protected beforeDestroy(): void {
+		this.memoizeImage();
+		this.$refs.img.style.backgroundImage = '';
+		super.beforeDestroy();
 	}
 }