+++ src/super/i-block/modules/state.ts
@@ -48,7 +48,7 @@ export default class State {
 	/**
 	 * Component instance
 	 */
-	protected readonly component: iBlock;
+	protected readonly component: iBlock['unsafe'];
 
 	/**
 	 * Component unique name
@@ -68,7 +68,6 @@ export default class State {
 	 * API for a component storage
 	 */
 	protected get storage(): Storage {
-		// @ts-ignore (access)
 		return this.component.storage;
 	}
 
@@ -76,7 +75,6 @@ export default class State {
 	 * API for lazy operations
 	 */
 	protected get lazy(): Lazy {
-		// @ts-ignore (access)
 		return this.component.lazy;
 	}
 
@@ -91,15 +89,13 @@ export default class State {
 	 * Local event emitter
 	 */
 	protected get localEvent(): Event {
-		// @ts-ignore (access)
 		return this.component.localEvent;
 	}
 
 	/**
 	 * Async instance
 	 */
-	protected get async(): Async {
-		// @ts-ignore (access)
+	protected get async(): Async<iBlock> {
 		return this.component.async;
 	}
 
@@ -109,8 +105,6 @@ export default class State {
 	protected get instance(): iBlock {
 		// @ts-ignore (access)
 		baseSyncRouterState = baseSyncRouterState || iBlock.prototype.syncRouterState;
-
-		// @ts-ignore (access)
 		return this.component.instance;
 	}
 
@@ -118,7 +112,7 @@ export default class State {
 	 * @param component - component instance
 	 */
 	constructor(component: iBlock) {
-		this.component = component;
+		this.component = component.unsafe;
 	}
 
 	/**
@@ -161,13 +155,10 @@ export default class State {
 		const
 			c = this.component;
 
-		// @ts-ignore (access)
 		data = c.syncStorageState(data, 'remote');
-		// @ts-ignore (access)
 		this.set(c.syncStorageState(data));
 
 		await this.storage.set(data, '[[STORE]]');
-		// @ts-ignore (access)
 		c.log('state:save:storage', this, data);
 
 		return true;
@@ -205,7 +196,6 @@ export default class State {
 
 			this.lfc.execCbAtTheRightTime(() => {
 				const
-					// @ts-ignore (access)
 					stateFields = c.syncStorageState(data);
 
 				this.set(
@@ -239,7 +229,6 @@ export default class State {
 					}
 				}
 
-				// @ts-ignore (access)
 				c.log('state:init:storage', this, stateFields);
 			});
 
@@ -265,7 +254,6 @@ export default class State {
 
 		const
 			c = this.component,
-			// @ts-ignore (access)
 			stateFields = c.convertStateToStorageReset();
 
 		this.set(
@@ -273,7 +261,6 @@ export default class State {
 		);
 
 		await this.saveToStorage();
-		// @ts-ignore (access)
 		c.log('state:reset:storage', this, stateFields);
 		return true;
 
@@ -294,9 +281,7 @@ export default class State {
 		const
 			c = this.component;
 
-		// @ts-ignore (access)
 		data = c.syncRouterState(data, 'remote');
-		// @ts-ignore (access)
 		this.set(c.syncRouterState(data));
 
 		const
@@ -310,7 +295,6 @@ export default class State {
 			query: data
 		});
 
-		// @ts-ignore (access)
 		c.log('state:save:router', this, data);
 		return true;
 
@@ -355,7 +339,6 @@ export default class State {
 
 			const
 				route = r.route || {},
-				// @ts-ignore (access)
 				stateFields = c.syncRouterState(Object.assign(Object.create(route), route.params, route.query));
 
 			this.set(
@@ -364,7 +347,6 @@ export default class State {
 
 			if (c.syncRouterStoreOnInit) {
 				const
-					// @ts-ignore (access)
 					routerState = c.syncRouterState(stateFields, 'remote');
 
 				if (Object.keys(routerState).length) {
@@ -416,7 +398,6 @@ export default class State {
 				}
 			}
 
-			// @ts-ignore (access)
 			c.log('state:init:router', this, stateFields);
 
 		}, {
@@ -440,7 +421,6 @@ export default class State {
 
 		const
 			c = this.component,
-			// @ts-ignore (access)
 			stateFields = c.convertStateToRouterReset();
 
 		this.set(
@@ -455,7 +435,6 @@ export default class State {
 		}
 
 		await r.push(null);
-		// @ts-ignore (access)
 		c.log('state:reset:router', this, stateFields);
 		return true;